refactor: 작품상세 뷰 앱바 객체화
